Add files via upload
Description: Develop a system for renting different types of vehicles.
●
Abstract Class: Vehicle with subclasses Car, Bike, and Truck.
●
Interface: Rentable with methods like rent() and returnVehicle().
●
Composition: RentalService class containing a list of Vehicle objects.
This project provides a basic structure for a Vehicle Rental System. It includes an abstract class Vehicle, subclasses Car, Bike, and Truck, an interface Rentable, and a RentalService class that uses composition to manage a list of vehicles.
Add a main class, the VehicleRentalSystem class, to see how it works to Test the System.